\name{as.POSIXcti}
\alias{as.POSIXcti}
\alias{as.POSIXcti.logical}
\title{Convert an object to a POSIXcti}
\usage{
  as.POSIXcti(from, ...)

  \method{as.POSIXcti}{logical} (from, ...)
}
\arguments{
  \item{from}{object to convert to a POSIXcti}

  \item{\dots}{more args to or from other methods}
}
\value{
  a \code{\link{POSIXcti}}
}
\description{
  Convert an object to a POSIXcti
}

