\name{as.POSIXctp}
\alias{as.POSIXctp}
\alias{as.POSIXctp.logical}
\title{Convert an object to a POSIXctp}
\usage{
  as.POSIXctp(from, ...)

  \method{as.POSIXctp}{logical} (from, ...)
}
\arguments{
  \item{from}{object to convert to a POSIXctp}

  \item{\dots}{more args to or from other methods}
}
\value{
  a \code{\link{POSIXctp}}
}
\description{
  Convert an object to a POSIXctp
}

