\name{\%included\%}
\alias{\%included\%}
\alias{\%included\%.POSIXcti}
\title{test inclusion of 2 \sQuote{POSIXcti} objects}
\usage{
  %included%(i1, i2)

  i1 \%included\% i2
}
\arguments{
  \item{i1}{is this object included in the second object ?}

  \item{i2}{is this object include the first one ?}
}
\value{
  boolean
}
\description{
  This function test if the first \sQuote{POSIXcti} object
  is included in the second.
}
\examples{
# to see all existing methods :
methods ('\\\%included\\\%')
}

