\name{\%intersect\%}
\alias{\%intersect\%}
\alias{\%intersect\%.POSIXcti}
\title{intersects 2 \sQuote{POSIXcti} objects}
\usage{
  %intersect%(i1, i2)

  i1 \%intersect\% i2
}
\arguments{
  \item{i1}{first object to intersect}

  \item{i2}{second object to intersect}
}
\value{
  object of the same class of parameters
}
\description{
  This function allows to find the intersection of two
  objects of the same class.
}
\examples{
# to see all existing methods :
methods ('\\\%intersect\\\%')
}

