\docType{data}
\name{origin}
\alias{origin}
\title{1970-01-01 GMT}
\format{POSIXt[1:1], format: "1970-01-01 01:00:00"}
\usage{
  origin
}
\description{
  Origin is the date-time for 1970-01-01 GMT in POSIXct
  format. This date-time is the origin for the numbering
  system used by POSIXct, POSIXlt, chron, and Date classes.
}
\details{
  The original implementation of this 'object' is in the
  lubridate package.
}
\examples{
origin
# "1970-01-01 GMT"
}
\author{
  Garrett Grolemund "grolemund at rice.edu", Hadley Wickham
  "h.wickham at gmail.com"
}
\keyword{chron}
\keyword{data}
\keyword{datasets}

