\name{ops.posixctpi}
\alias{-}
\alias{*}
\alias{+}
\alias{*,numeric,POSIXctp-method}
\alias{Ops.POSIXcti}
\alias{Ops.POSIXctp}
\alias{ops.posixctpi}
\alias{-,POSIXcti,POSIXctp-method}
\alias{+,POSIXcti,POSIXctp-method}
\alias{-,POSIXct,POSIXctp-method}
\alias{+,POSIXct,POSIXctp-method}
\alias{+,POSIXctp,POSIXcti-method}
\alias{+,POSIXctp,POSIXct-method}
\alias{+,POSIXctp,POSIXctp-method}
\title{Operators to sum or substract POSIXctp from POSIXct or POSIXcti}
\description{
  Operators to sum or substract POSIXctp from POSIXct or
  POSIXcti
}

