\name{POSIXt.units}
\alias{POSIXt.units}
\title{define valid units for time objects}
\usage{
  POSIXt.units(x = NULL, ...)
}
\arguments{
  \item{x}{a character string representing the needed
  units}

  \item{\dots}{arguments to or from other methods}
}
\description{
  With no argument, the functions return a factor
  containing the valid time units. With an argument, it
  returns the units asked for.
}

